requirements.txt
click==8.1.7
anthropic==0.45.0
rich==13.9.4
python-dotenv==1.0.1
pylint==3.3.2
pyyaml==6.0.2
# semgrep==1.100.0  # Windows 미지원 - WSL/Linux 환경에서만 설치 가능

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
